using unique user name + using pre-gen config/key

+++ src/bootstrap.sh
@@ -2,27 +2,44 @@
 
 #Installing docker
 yum update -y
-amazon-linux-extras -y install docker
 yum install -y docker conntrack
 service docker start
 
 confdir=/tmp/openvpn
 ovpnport=1194
 domain="DOMAIN_NAME" #TBU
-user=user1
+user=USERNAME
 dimage=damshenas/openvpn:arm64
 
-# preparing the keys and running the openvpn server
 mkdir -p $confdir
+aws s3 sync s3://ARTIFACTS_S3_BUCKET/openvpn $confdir
+
+# preparing the keys and running the openvpn server
 docker pull $dimage
-docker run -v $confdir:/etc/openvpn --rm $dimage ovpn_genconfig -u udp://$domain
-docker run -v $confdir:/etc/openvpn --rm -i -e "EASYRSA_BATCH=1" -e "EASYRSA_REQ_CN=My CN" $dimage ovpn_initpki nopass
+
+if [ ! -f "$confdir/openvpn.conf" ]; then
+  echo "Config not found. Generating config."
+  docker run -v $confdir:/etc/openvpn --rm $dimage ovpn_genconfig -u udp://$domain
+fi
+
+if [ ! -f "$confdir/pki/ca.crt" ]; then
+  echo "Cert not found. Generating certificates."
+  docker run -v $confdir:/etc/openvpn --rm -i -e "EASYRSA_BATCH=1" -e "EASYRSA_REQ_CN=My CN" $dimage ovpn_initpki nopass
+fi
+
 docker run -v $confdir:/etc/openvpn -d -p $ovpnport:$ovpnport/udp --cap-add=NET_ADMIN $dimage
+
+if [ ! -f "$confdir/pki/reqs/$user.req" ]; then
+  rm -f $confdir/pki/reqs/$user.req
+  rm -f $confdir/pki/issued/$user.req
+  rm -f $confdir/pki/private/$user.req
+  rm -f $confdir/pki/reqs/$user.req
+fi
+
+# create user cert and profile. And upload to s3.
 docker run -v $confdir:/etc/openvpn --rm -i $dimage easyrsa build-client-full $user nopass
 docker run -v $confdir:/etc/openvpn --rm $dimage ovpn_getclient $user > $user.ovpn
 docker run -v $confdir:/etc/openvpn --rm $dimage ovpn_listclients | grep $user
-
-# upload the openvpn config to s3
 aws s3 cp $user.ovpn s3://ARTIFACTS_S3_BUCKET/profiles/$user.ovpn
 
 # minitor connections
@@ -32,7 +49,7 @@ max_minutes_without_connection=15
 while true
 do
   # no_connections=$(ss -tun src :$ovpnport | grep ESTAB | wc -l)
-  # no_users=$(docker exec -it intelligent_mclean cat /tmp/openvpn-status.log | grep user1)
+  # no_users=$(docker exec -it intelligent_mclean cat /tmp/openvpn-status.log | grep $user)
   # no_connections=$(conntrack -L --proto udp --dport 1194 --status ASSURED)
   no_connections=$(cat /proc/net/nf_conntrack | grep ASSURED | grep 1194 | wc -l)
